+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changes
 
+- Unreleased
+  * Add support for Django's `modelform_factory` and `Meta.fields` option.
+  * Add support for Django 5.0, 5.1
+  * Drop support for Django 4.1, 4.0, 3.2, 2.2
+
 - 0.5.4
   * Add support for Django-4.1 and 4.2.
   * Confirm support for Python 3.11